Renamed folder to clicks
